backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-decouple==3.8
python-dotenv==1.0.0
google-generativeai==0.3.2
reportlab==4.0.7